+++ visualization/app/codeCharta/ui/metricChooser/metricType/isHoveredNodeALeaf.selector.ts
@@ -1,11 +1,7 @@
 import { createSelector } from "../../../state/angular-redux/createSelector"
-import { idToNodeSelector } from "../../../state/selectors/accumulatedData/idToNode.selector"
-import { hoveredNodeIdSelector } from "../../../state/store/appStatus/hoveredNodeId/hoveredNodeId.selector"
-import { isLeaf, MaybeLeaf } from "../../../util/codeMapHelper"
+import { hoveredNodeSelector } from "../../../state/selectors/hoveredNode.selector"
+import { isLeaf } from "../../../util/codeMapHelper"
 
-export const _isHoveredNodeALeaf = (idToNode: Map<number, MaybeLeaf>, hoveredNodeId: number | null) => {
-	const node = idToNode.get(hoveredNodeId)
-	return node && isLeaf(node)
-}
+export const _isHoveredNodeALeaf = hoveredNode => hoveredNode && isLeaf(hoveredNode)
 
-export const isHoveredNodeALeafSelector = createSelector([idToNodeSelector, hoveredNodeIdSelector], _isHoveredNodeALeaf)
+export const isHoveredNodeALeafSelector = createSelector([hoveredNodeSelector], _isHoveredNodeALeaf)