+++ visualization/app/codeCharta/ui/metricChooser/metricValueHovered/metricValueHovered.module.ts
@@ -1,0 +1,12 @@
+import { CommonModule } from "@angular/common"
+import { NgModule } from "@angular/core"
+import { MetricTypeModule } from "../metricType/metricType.module"
+import { MetricValueHoveredComponent } from "./metricValueHovered.component"
+
+@NgModule({
+	imports: [CommonModule, MetricTypeModule],
+	declarations: [MetricValueHoveredComponent],
+	exports: [MetricValueHoveredComponent],
+	entryComponents: [MetricValueHoveredComponent]
+})
+export class MetricValueHoveredModule {}